+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `Cryo<T, _>: Send` requires `T: Sync` instead of `T: Send`.
     - `CryoMut<T, _>: Send` now requires `T: Sync` in addition to `T: Send`.
     - `CryoMutReadGuard<T, _>: Send` requires `T: Sync` instead of `T: Send`.
+- Breaking: The `Send`/`Sync`-ness of `Lock` is now taken into account in trait bounds.
 - `Cryo<T, _>: Sync` no longer requires `T: Send`.
 - `CryoMutReadGuard<T, _>: Sync` no longer requires `T: Send`.
 - `CryoMutWriteGuard<T, _>: Sync` if `T: Sync`.